+++ packages/insight/src/Access/DiagnosticAccessOperation.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Insight\Access;
+
+enum DiagnosticAccessOperation: string
+{
+    case List = 'list';
+    case Detail = 'detail';
+}